Add comments; remove $extractions param from get-existing-match-notification
I removed the $extractions parameter because the way that function is
used, the extractions are never relevant. They only matter when the
result is getting passed back to the user. The get-existing-match-notification
function is an internal-only utility function. Every call to it was
passing in an empty map, except for a unit test (notifications-to-json).
I changed that unit test to call getNotifications and then pick the one
it wanted to test.